Fix tests and formatting
