+++ CHANGES.md
@@ -10,7 +10,7 @@
   decorator. If specified, it defines an _inputs argument_, which is used to 
   define the process inputs in form of a dataclass derived from 
   `pydantic.BaseModel`. (#35)
-- Now supporting the OpenAPI/JSON Schema `disscriminator` property 
+- Now supporting the OpenAPI/JSON Schema `discriminator` property 
   in `tools/openapi.yaml/schemas/schema` and generated models. (#36)
 - Renamed `s2gos_common.process.cli.ProcessingRequest` into
   `s2gos_common.process.cli.ExecutionRequest`.
