+++ s2gos-app-ex/src/s2gos_app_ex/processes.py
@@ -5,6 +5,7 @@
 from typing import Annotated
 
 from pydantic import Field
+
 from s2gos_common.process import JobContext, ProcessRegistry
 
 registry = ProcessRegistry()
