+++ s2gos-server/src/s2gos_server/app.py
@@ -45,20 +45,3 @@ async def log_request_duration(
     )
 
     return response
-
-
-class LogMessageFilter(logging.Filter):
-    def __init__(self, *excludes: str):
-        super().__init__(f"Log message filter: {excludes}")
-        self.excludes = excludes
-
-    def filter(self, record: logging.LogRecord) -> bool:
-        if record.name == "uvicorn.access":
-            for exclude in self.excludes:
-                if exclude in record.getMessage():
-                    return False
-        return True
-
-
-# Apply the filter to the uvicorn.access logger
-logging.getLogger("uvicorn.access").addFilter(LogMessageFilter("/jobs"))
