+++ tools/gen_dags.py
@@ -8,8 +8,8 @@ from typing import Any
 
 import typer
 
-from s2gos_server.services.local import LocalService
 from s2gos_common.process import Process
+from s2gos_server.services.local import LocalService
 from tools.common import S2GOS_PATH, write_file
 
 GENERATOR_NAME = str(Path(__file__).name)
